migrate to navigation component
