+++ app/src/main/res/values/strings.xml
@@ -18,4 +18,6 @@
     <string name="show_today">Show today\'s weather</string>
 
     <string name="search_city">Enter your city here</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
